+++ km3modules/ahrs.py
@@ -8,12 +8,15 @@ AHRS calibration.
 from __future__ import division, absolute_import, print_function
 
 import io
+from collections import defaultdict
+import time
 from functools import lru_cache
 import xml.etree.ElementTree as ET
 
 import numpy as np
 from numpy import cos, sin, arctan2
 import km3pipe as kp
+from km3pipe.io.daq import TMCHData
 
 __author__ = "Tamas Gal"
 __email__ = "tgal@km3net.de"
@@ -21,6 +24,84 @@ __status__ = "Development"
 
 log = kp.logger.logging.getLogger(__name__)  # pylint: disable=C0103
 # log.setLevel(logging.DEBUG)
+
+
+class AHRSCalibrator(kp.Module):
+    """Calculates AHRS yaw, pitch and roll from median A and H of an interval.
+
+    Required Parameters
+    -------------------
+    det_id: int (detector ID, e.g. 29)
+
+    Optional Parameters
+    -------------------
+    interval: int (accumulation interval in [sec], default: 10s)
+
+    Output Keys
+    -----------
+    'AHRSCalibration':
+        dict: key=dom_id, value=tuple: (timestamp, du, floor, yaw, pitch, roll)
+
+    """
+    def configure(self):
+        det_id = self.require('det_id')
+        self.interval = self.get('interval') or 10  # in sec
+        self.A = defaultdict(list)
+        self.H = defaultdict(list)
+        self.detector = kp.hardware.Detector(det_id=det_id)
+        self.db = self.get('dbmanager') or kp.db.DBManager()
+        self.timestamp = time.time()
+
+    def process(self, blob):
+        tmch_data = TMCHData(io.BytesIO(blob['CHData']))
+        dom_id = tmch_data.dom_id
+        try:
+            du, floor, _ = self.detector.doms[dom_id]
+        except KeyError:  # base CLB
+            return blob
+
+        self.A[dom_id].append(tmch_data.A)
+        self.H[dom_id].append(tmch_data.H)
+
+        if time.time() - self.timestamp > self.interval:
+            self.timestamp = time.time()
+            calib = self.calibrate()
+            blob['AHRSCalibration'] = calib
+
+        return blob
+
+    def calibrate(self):
+        """Calculate yaw, pitch and roll from the median of A and H.
+
+        After successful calibration, the `self.A` and `self.H` are reset.
+        DOMs with missing AHRS pre-calibration data are skipped.
+
+        Returns
+        -------
+        dict: key=dom_id, value=tuple: (timestamp, du, floor, yaw, pitch, roll)
+
+        """
+        now = time.time()
+        dom_ids = self.A.keys()
+        print("Calibrating AHRS from median A and H for {} DOMs."
+              .format(len(dom_ids)))
+        calibrations = {}
+        for dom_id in dom_ids:
+            print("Calibrating DOM ID {}".format(dom_id))
+            clb_upi = self.db.doms.via_dom_id(dom_id).clb_upi
+            ahrs_calib = get_latest_ahrs_calibration(clb_upi)
+            if ahrs_calib is None:
+                log.warn("No AHRS calibration found for '{}'".format(dom_id))
+                continue
+            du, floor, _ = self.detector.doms[dom_id]
+            A = np.median(self.A[dom_id], axis=0)
+            H = np.median(self.H[dom_id], axis=0)
+
+            cyaw, cpitch, croll = fit_ahrs(A, H, *ahrs_calib)
+            calibrations[dom_id] = (now, du, floor, cyaw, cpitch, croll)
+        self.A = defaultdict(list)
+        self.H = defaultdict(list)
+        return calibrations
 
 
 def fit_ahrs(A, H, Aoff, Arot, Hoff, Hrot):
@@ -45,19 +126,19 @@ def fit_ahrs(A, H, Aoff, Arot, Hoff, Hrot):
     Acal = np.dot(A-Aoff, Arot)
     Hcal = np.dot(H-Hoff, Hrot)
 
-    #invert axis for DOM upside down
+    # invert axis for DOM upside down
     for i in (1, 2):
         Acal[i] = -Acal[i]
         Hcal[i] = -Hcal[i]
 
     roll = arctan2(-Acal[1], -Acal[2])
-    pitch= arctan2(Acal[0], np.sqrt(Acal[1]*Acal[1] + Acal[2]*Acal[2]))
+    pitch = arctan2(Acal[0], np.sqrt(Acal[1]*Acal[1] + Acal[2]*Acal[2]))
     yaw = arctan2(Hcal[2]*sin(roll) - Hcal[1]*cos(roll),
                   sum((Hcal[0]*cos(pitch),
                        Hcal[1]*sin(pitch)*sin(roll),
                        Hcal[2]*sin(pitch)*cos(roll))))
 
-    #yaw = (yaw + magnetic_declination + 360 ) % 360
+    # yaw = (yaw + magnetic_declination + 360 ) % 360
     yaw = np.degrees(yaw)
     roll = np.degrees(roll)
     pitch = np.degrees(pitch)
@@ -138,12 +219,11 @@ def _extract_calibration(xroot):
     Arot_ix = names.index("AHRS_Acceleration_Rotation(-)")
     Hrot_ix = names.index("AHRS_Magnetic_Rotation(-)")
 
-    Aoff = np.array(val[Aoff_ix])[vec_ic]  \
-           .astype(float)
-    Arot = np.array(val[Arot_ix]).reshape(3, 3)[col_ic, row_ic].reshape(3, 3) \
-           .astype(float)
-    Hrot = np.array(val[Hrot_ix]).reshape(3, 3)[col_ic, row_ic].reshape(3, 3) \
-           .astype(float)
+    Aoff = np.array(val[Aoff_ix])[vec_ic].astype(float)
+    Arot = np.array(val[Arot_ix]).reshape(3, 3)[col_ic, row_ic] \
+                                 .reshape(3, 3).astype(float)
+    Hrot = np.array(val[Hrot_ix]).reshape(3, 3)[col_ic, row_ic] \
+                                 .reshape(3, 3).astype(float)
 
     Hoff = []
     for q in 'XYZ':