+++ km3pipe/io/jpp.py
@@ -13,7 +13,7 @@ import numpy as np
 from km3pipe.core import Pump, Blob
 from km3pipe.dataclasses import (EventInfo, TimesliceFrameInfo,
                                  SummaryframeInfo, HitSeries,
-                                 TimesliceHitSeries)
+                                 TimesliceHitSeries, RawHitSeries)
 from km3pipe.logger import logging
 
 log = logging.getLogger(__name__)  # pylint: disable=C0103
@@ -29,6 +29,9 @@ __status__ = "Development"
 
 class JPPPump(Pump):
     """A pump for JPP ROOT files.
+    
+    This pump will be replaced soon by ``EventPump``, ``TimeslicePump`` and
+    ``SummaryslicePump``.
 
     Parameters
     ----------
@@ -192,6 +195,204 @@ class JPPPump(Pump):
 
     def process(self, blob):
         return next(self.blobs)
+
+    def __iter__(self):
+        return self
+
+    def next(self):
+        """Python 2/3 compatibility for iterators"""
+        return self.__next__()
+
+    def __next__(self):
+        return next(self.blobs)
+
+
+class EventPump(Pump):
+    """A pump for DAQEvents in JPP files.
+    
+    Parameters
+    ----------
+    filename: str
+        Name of the file to open. 
+
+    """
+    def configure(self):
+
+        try:
+            import jppy  # noqa
+        except ImportError:
+            raise ImportError("\nPlease install the jppy package:\n\n"
+                              "    pip install jppy\n")
+
+        self.event_index = self.get('index') or 0
+        self.filename = self.require('filename')
+
+        self.event_reader = jppy.PyJDAQEventReader(self.filename)
+        self.blobs = self.blob_generator()
+
+    def blob_generator(self):
+        while self.event_reader.has_next:
+            try:
+                yield self.extract_event()
+            except IndexError:
+                pass
+
+        raise StopIteration
+
+    def extract_event(self):
+        blob = Blob()
+        r = self.event_reader
+        r.retrieve_next_event()  # do it at the beginning!
+
+        n = r.number_of_snapshot_hits
+        channel_ids = np.zeros(n, dtype='i')
+        dom_ids = np.zeros(n, dtype='i')
+        times = np.zeros(n, dtype='i')
+        tots = np.zeros(n, dtype='i')
+        triggereds = np.zeros(n, dtype='i')
+
+        r.get_hits(channel_ids, dom_ids, times, tots, triggereds)
+
+        nans = np.full(n, np.nan, dtype='<f8')
+        hit_series = HitSeries.from_arrays(
+            channel_ids, nans, nans, nans, dom_ids, np.arange(n), np.zeros(n),
+            nans, nans, nans, nans, times, tots, triggereds, self.event_index
+        )
+
+        event_info = EventInfo(np.array((
+            r.det_id, r.frame_index,
+            0,  # livetime_sec
+            0, 0,  # MC ID and time
+            0,  # n_events_gen
+            0,  # n_files_gen
+            r.overlays,
+            r.trigger_counter, r.trigger_mask,
+            r.utc_nanoseconds, r.utc_seconds,
+            np.nan, np.nan, np.nan,   # w1-w3
+            0,  # run
+            self.event_index,
+            ), dtype=EventInfo.dtype))
+
+        self.event_index += 1
+        blob['EventInfo'] = event_info
+        blob['Hits'] = hit_series
+        return blob
+
+    def process(self, blob):
+        return next(self.blobs)
+
+    def __iter__(self):
+        return self
+
+    def next(self):
+        """Python 2/3 compatibility for iterators"""
+        return self.__next__()
+
+    def __next__(self):
+        return next(self.blobs)
+
+
+class TimeslicePump(Pump):
+    """A pump to read and extract timeslices. Currently only hits are read.
+    
+    Required Parameters
+    -------------------
+    filename: str
+    
+    """
+    def configure(self):
+        filename = self.require('filename')
+        self.blobs = self.timeslice_generator()
+        try:
+            import jppy  # noqa
+        except ImportError:
+            raise ImportError("\nPlease install the jppy package:\n\n"
+                              "    pip install jppy\n")
+        self.r = jppy.daqtimeslicereader.PyJDAQTimesliceReader(filename)
+
+    def process(self, blob):
+        blob['TSHits'] = next(self.blobs)
+        return blob
+
+    def timeslice_generator(self):
+        while self.r.has_next:
+            slice_id = 1
+            self.r.retrieve_next_timeslice()
+            channel_ids = np.array(())
+            dom_ids = np.array(())
+            times = np.array(())
+            tots = np.array(())
+            n_frames = 0
+            while self.r.has_next_superframe:
+                n_frames += 1
+                n = self.r.number_of_hits
+                if n == 0:
+                    self.r.retrieve_next_superframe()
+                    continue
+                _channel_ids = np.zeros(n, dtype='i')
+                _dom_ids = np.zeros(n, dtype='i')
+                _times = np.zeros(n, dtype='i')
+                _tots = np.zeros(n, dtype='i')
+                self.r.get_hits(_channel_ids, _dom_ids, _times, _tots)
+                channel_ids = np.concatenate((channel_ids, _channel_ids))
+                dom_ids = np.concatenate((dom_ids, _dom_ids))
+                times = np.concatenate((times, _times))
+                tots = np.concatenate((tots, _tots))
+                self.r.retrieve_next_superframe()
+
+            hits = RawHitSeries.from_arrays(
+                    channel_ids, dom_ids, times, tots,
+                    np.zeros(len(channel_ids), dtype=bool), slice_id)
+            yield hits
+            slice_id += 1
+
+    def __iter__(self):
+        return self
+
+    def next(self):
+        """Python 2/3 compatibility for iterators"""
+        return self.__next__()
+
+    def __next__(self):
+        return next(self.blobs)
+
+
+class SummaryslicePump(Pump):
+    """Preliminary Summaryslice reader"""
+    def configure(self):
+        filename = self.require('filename')
+        self.blobs = self.summaryslice_generator()
+        try:
+            import jppy  # noqa
+        except ImportError:
+            raise ImportError("\nPlease install the jppy package:\n\n"
+                              "    pip install jppy\n")
+        self.r = jppy.daqsummaryslicereader.PyJDAQSummarysliceReader(filename)
+
+    def process(self, blob):
+        blob['Summaryslice'] = next(self.blobs)
+        return blob
+
+    def summaryslice_generator(self):
+        while self.r.has_next:
+            summary_slice = {}
+            self.r.retrieve_next_summaryslice()
+            while self.r.has_next_frame:
+                rates = np.zeros(31, dtype='f8')
+                self.r.get_rates(rates)
+                summary_slice[self.r.dom_id] = {
+                        'rates': rates,
+                        'n_udp_packets': self.r.number_of_received_packets,
+                        'max_sequence_number': self.r.max_sequence_number,
+                        'has_udp_trailer': self.r.has_udp_trailer,
+                        'high_rate_veto': self.r.high_rate_veto,
+                        'fifo_status': self.r.fifo_status,
+                        'frame_index': self.r.frame_index,
+                        'utc_seconds': self.r.utc_seconds,
+                        'utc_nanoseconds': self.r.utc_nanoseconds,
+                        }
+                self.r.retrieve_next_frame()
+            yield summary_slice
 
     def __iter__(self):
         return self